costs seem accurate now

+++ algolabra/fileIO/read_files.py
@@ -1,3 +1,6 @@
+from decimal import Decimal
+
+
 def read_map(filename) -> list:
     """
     Read movingai-type map file and preprocess it.
@@ -17,7 +20,7 @@ def read_scenarios(filename):
     map_title = lines[0][1]
     # bucket, map, (start), (goal), ideal_solution
     converted = [[int(a[0]), (int(a[4]), int(a[5])), (int(a[6]), int(a[7])),
-                  float(a[8])] for a in lines]
+                  Decimal(a[8])] for a in lines]
     for i, scenario in enumerate(converted):
         new_item = [i, *scenario]
         if scenario[0] in product:
